+++ packages/core/src/react/hooks/__tests__/useFetchPosts.tsx
@@ -2,18 +2,27 @@ import { renderHook, waitFor } from '@testing-library/react';
 import { expectTypeOf } from 'expect-type';
 import * as React from 'react';
 
+import { SWRConfig } from 'swr';
 import { PostEntity, PostsArchiveParams } from '../../../data';
 import { SettingsProvider } from '../../provider';
 import { useFetchPosts } from '../useFetchPosts';
 import { setHeadlessConfig } from '../../../utils';
+import * as useFetchModule from '../useFetch';
+import { mockUseFetchErrorResponse } from '../mocks';
 
 describe('useFetchPosts', () => {
 	const wrapper = ({ children }) => {
-		return <SettingsProvider settings={{ sourceUrl: '' }}>{children}</SettingsProvider>;
+		return (
+			<SWRConfig value={{ provider: () => new Map() }}>
+				<SettingsProvider settings={{ sourceUrl: '' }}>{children}</SettingsProvider>
+			</SWRConfig>
+		);
 	};
 
-	setHeadlessConfig({
-		useWordPressPlugin: true,
+	beforeEach(() => {
+		setHeadlessConfig({
+			useWordPressPlugin: true,
+		});
 	});
 
 	it('throws errors if accessing data before fetch', async () => {
@@ -38,6 +47,33 @@ describe('useFetchPosts', () => {
 		await waitFor(() => {
 			expect(result.current.data?.posts.length).toBe(2);
 		});
+	});
+
+	it('handles response if has error or there is no data', async () => {
+		const spyUseFetch = jest
+			.spyOn(useFetchModule, 'useFetch')
+			.mockReturnValueOnce(mockUseFetchErrorResponse);
+		const { result } = renderHook(() => useFetchPosts({}), {
+			wrapper,
+		});
+
+		const expectedKeys = ['error', 'loading', 'pageType', 'data', 'isMainQuery'];
+		const returnedKeys = Object.keys(result.current);
+		const missingKeys = returnedKeys.filter((key) => !expectedKeys.includes(key));
+
+		await waitFor(() => {
+			expect(missingKeys).toHaveLength(0);
+			expect(spyUseFetch).toHaveBeenCalledTimes(1);
+			expect(result.current.error).toBe('Not found');
+			expect(result.current.loading).toBe(true);
+			expect(() => result.current.data).not.toThrow();
+			expect(() => result.current.data?.posts[0].title).toThrow();
+			expect(() => result.current.data?.pageInfo[0].title).toThrow();
+			expect(() => result.current.data?.queriedObject[0].title).toThrow();
+			expect(result.current.isMainQuery).toBe(true);
+		});
+
+		spyUseFetch.mockRestore();
 	});
 
 	it('returns queried object for category archives', async () => {
@@ -235,24 +271,114 @@ describe('useFetchPosts', () => {
 		});
 	});
 
-	describe('useFetchPosts types', () => {
-		it('allows overriding types', () => {
-			interface Book extends PostEntity {
-				isbn: string;
-			}
+	it('throws params.matchArchivepath is true and path does not match', async () => {
+		const { result } = renderHook(
+			() =>
+				useFetchPosts(
+					{
+						category: 'uncategorized',
+						per_page: 1,
+						matchArchivePath: true,
+					},
+					{},
+					'https://js1.10up.com/category/asdasd/uncategorized',
+				),
+			{
+				wrapper,
+			},
+		);
 
-			interface BookParams extends PostsArchiveParams {
-				isbn: string;
-			}
-
-			const { result } = renderHook(() => useFetchPosts<Book, BookParams>({ isbn: 'sdasd' }));
-
-			expectTypeOf(result.current.data?.posts).toMatchTypeOf<
-				| Array<{
-						isbn: string;
-				  }>
-				| undefined
-			>();
+		await waitFor(() => {
+			expect(result.current.error?.toString()).toBe(
+				`NotFoundError: Posts were found but did not match current path: "https://js1.10up.com/category/asdasd/uncategorized"`,
+			);
 		});
+	});
+
+	it('throws matchArchivepath config option is true and path does not match', async () => {
+		setHeadlessConfig({
+			useWordPressPlugin: true,
+			customTaxonomies: (defaultTaxonomies) => {
+				return defaultTaxonomies.map((taxonomy) => ({
+					...taxonomy,
+					matchArchivePath: true,
+				}));
+			},
+		});
+
+		const { result } = renderHook(
+			() =>
+				useFetchPosts(
+					{
+						category: 'uncategorized',
+						per_page: 2,
+					},
+					{},
+					'https://js1.10up.com/category/asdasd/uncategorized',
+				),
+			{
+				wrapper,
+			},
+		);
+
+		await waitFor(() => {
+			expect(result.current.error?.toString()).toBe(
+				`NotFoundError: Posts were found but did not match current path: "https://js1.10up.com/category/asdasd/uncategorized"`,
+			);
+		});
+	});
+
+	it('does not throws when matchArchivepath config option is true and path matches', async () => {
+		setHeadlessConfig({
+			useWordPressPlugin: true,
+			customTaxonomies: (defaultTaxonomies) => {
+				return defaultTaxonomies.map((taxonomy) => ({
+					...taxonomy,
+					matchArchivePath: true,
+				}));
+			},
+		});
+
+		const { result } = renderHook(
+			() =>
+				useFetchPosts(
+					{
+						// randomArgument: 10, // bypass swr cache
+						category: 'uncategorized',
+						per_page: 1,
+					},
+					{},
+					// Need this bc source url removal is not working in the tests
+					'https://js1.10up.com/category/uncategorized',
+				),
+			{
+				wrapper,
+			},
+		);
+
+		await waitFor(() => {
+			expect(result.current.data?.queriedObject.term?.slug).toBe('uncategorized');
+		});
+	});
+});
+
+describe('useFetchPosts types', () => {
+	it('allows overriding types', () => {
+		interface Book extends PostEntity {
+			isbn: string;
+		}
+
+		interface BookParams extends PostsArchiveParams {
+			isbn: string;
+		}
+
+		const { result } = renderHook(() => useFetchPosts<Book, BookParams>({ isbn: 'sdasd' }));
+
+		expectTypeOf(result.current.data?.posts).toMatchTypeOf<
+			| Array<{
+					isbn: string;
+			  }>
+			| undefined
+		>();
 	});
 });