+++ packages/core/src/react/hooks/mocks.ts
@@ -1,0 +1,9 @@
+export const mockUseFetchErrorResponse = {
+	error: 'Not found',
+	params: {},
+	data: undefined,
+	isMainQuery: true,
+	mutate: jest.fn(),
+	isLoading: false,
+	isValidating: false,
+};